Moving input reader into run_local.
